requirements.txt
bs4
PyYAML